--- a/lmdb/lmdb.go
+++ b/lmdb/lmdb.go
@@ -59,15 +59,9 @@ package lmdb
 #cgo openbsd CFLAGS: -DMDB_DSYNC=O_SYNC
 #cgo netbsd CFLAGS: -DMDB_DSYNC=O_SYNC
 
-#include <stdlib.h>
 #include "lmdb.h"
-#include "lmdbgo.h"
 */
 import "C"
-import (
-	"sync"
-	"unsafe"
-)
 
 // Version return the major, minor, and patch version numbers of the LMDB C
 // library and a string representation of the version.
@@ -91,82 +85,6 @@ func VersionString() string {
 func cbool(b bool) C.int {
 	if b {
 		return 1
-	}
-	return 0
-}
-
-// msgctx is the type used for context pointers passed to mdb_reader_list.  A
-// msgctx stores its corresponding msgfunc, and any error encountered in an
-// external map.  The corresponding function is called once for each
-// mdb_reader_list entry using the msgctx.
-//
-// External maps are required because struct pointers passed to C functions
-// must not contain pointers in their struct fields.  See the following
-// language proposal which discusses the restrictions on passing pointers to C.
-//
-//		https://github.com/golang/proposal/blob/master/design/12416-cgo-pointers.md
-//
-// NOTE:
-// The underlying type must have a non-zero size to ensure that the value
-// returned by new(msgctx) does not conflict with other live *msgctx values.
-type msgctx byte
-type msgfunc func(string) error
-
-var msgctxm = map[*msgctx]msgfunc{}
-var msgctxe = map[*msgctx]error{}
-var msgctxmlock sync.RWMutex
-
-func registerMsgctx(fn msgfunc) *msgctx {
-	ctx := new(msgctx)
-	ctx.register(fn)
-	return ctx
-}
-
-func (ctx *msgctx) register(fn msgfunc) {
-	msgctxmlock.Lock()
-	msgctxm[ctx] = fn
-	msgctxmlock.Unlock()
-}
-
-func (ctx *msgctx) deregister() {
-	msgctxmlock.Lock()
-	delete(msgctxm, ctx)
-	delete(msgctxe, ctx)
-	msgctxmlock.Unlock()
-}
-
-func (ctx *msgctx) fn() msgfunc {
-	msgctxmlock.Lock()
-	fn := msgctxm[ctx]
-	msgctxmlock.Unlock()
-	return fn
-}
-
-func (ctx *msgctx) err() error {
-	msgctxmlock.Lock()
-	err := msgctxe[ctx]
-	msgctxmlock.Unlock()
-	return err
-}
-
-func (ctx *msgctx) seterr(err error) {
-	msgctxmlock.Lock()
-	msgctxe[ctx] = err
-	msgctxmlock.Unlock()
-}
-
-//export lmdbgoMDBMsgFuncBridge
-func lmdbgoMDBMsgFuncBridge(msg C.lmdbgo_ConstCString, _ctx unsafe.Pointer) C.int {
-	ctx := (*msgctx)(_ctx)
-	fn := ctx.fn()
-	if fn == nil {
-		return 0
-	}
-
-	err := fn(C.GoString(msg.p))
-	if err != nil {
-		ctx.seterr(err)
-		return -1
 	}
 	return 0
 }