--- a/lmdb/lmdbgo.c
+++ b/lmdb/lmdbgo.c
@@ -9,7 +9,7 @@ int lmdbgo_mdb_msg_func_proxy(const char *msg, void *ctx) {
     //  wrap msg and call the bridge function exported from lmdb.go.
     lmdbgo_ConstCString s;
     s.p = msg;
-    return lmdbgoMDBMsgFuncBridge(s, ctx);
+    return lmdbgoMDBMsgFuncBridge(s, (size_t)ctx);
 }
 
 int lmdbgo_mdb_reader_list(MDB_env *env, size_t ctx) {